requirements.txt
fastapi
uvicorn
pydantic
openai
pyyaml
openenv-core>=0.2.0
